+++ pylatex/document.py
@@ -7,10 +7,11 @@ This module implements the class that deals with the full document.
 """
 
 import os
+import sys
 import subprocess
 import errno
 from .base_classes import Environment, Command, Container, LatexObject, \
-    UnsafeCommand
+    UnsafeCommand, SpecialArguments
 from .package import Package
 from .errors import CompilerError
 from .utils import dumps_list, rm_temp_dir, NoEscape
@@ -60,7 +61,7 @@ class Document(Environment):
             Determines whether or not the document requires indentation. If it
             is `None` it will use the value from the active config. Which is
             `True` by default.
-        geometry_options: str or list
+        geometry_options: dict
             The options to supply to the geometry package
         data: list
             Initial content of the document.
@@ -104,7 +105,14 @@ class Document(Environment):
             packages.append(Package('microtype'))
 
         if geometry_options is not None:
-            packages.append(Package('geometry', options=geometry_options))
+            packages.append(Package('geometry'))
+            # Make sure we don't add this options command for an empty list,
+            # because that breaks.
+            if geometry_options:
+                packages.append(Command(
+                    'geometry',
+                    arguments=SpecialArguments(geometry_options),
+                ))
 
         super().__init__(data=data)
 
@@ -197,19 +205,25 @@ class Document(Environment):
         if compiler_args is None:
             compiler_args = []
 
+        # In case of newer python with the use of the cwd parameter
+        # one can avoid to physically change the directory
+        # to the destination folder
+        python_cwd_available = sys.version_info >= (3, 6)
+
         filepath = self._select_filepath(filepath)
-        filepath = os.path.join('.', filepath)
+        if not os.path.basename(filepath):
+            filepath = os.path.join(os.path.abspath(filepath),
+                                    'default_basename')
+        else:
+            filepath = os.path.abspath(filepath)
 
         cur_dir = os.getcwd()
         dest_dir = os.path.dirname(filepath)
-        basename = os.path.basename(filepath)
 
-        if basename == '':
-            basename = 'default_basename'
+        if not python_cwd_available:
+            os.chdir(dest_dir)
 
-        os.chdir(dest_dir)
-
-        self.generate_tex(basename)
+        self.generate_tex(filepath)
 
         if compiler is not None:
             compilers = ((compiler, []),)
@@ -221,7 +235,11 @@ class Document(Environment):
                 ('pdflatex', [])
             )
 
-        main_arguments = ['--interaction=nonstopmode', basename + '.tex']
+        main_arguments = ['--interaction=nonstopmode', filepath + '.tex']
+
+        check_output_kwargs = {}
+        if python_cwd_available:
+            check_output_kwargs = {'cwd': dest_dir}
 
         os_error = None
 
@@ -230,7 +248,8 @@ class Document(Environment):
 
             try:
                 output = subprocess.check_output(command,
-                                                 stderr=subprocess.STDOUT)
+                                                 stderr=subprocess.STDOUT,
+                                                 **check_output_kwargs)
             except (OSError, IOError) as e:
                 # Use FileNotFoundError when python 2 is dropped
                 os_error = e
@@ -250,16 +269,17 @@ class Document(Environment):
             if clean:
                 try:
                     # Try latexmk cleaning first
-                    subprocess.check_output(['latexmk', '-c', basename],
-                                            stderr=subprocess.STDOUT)
-                except (OSError, IOError, subprocess.CalledProcessError) as e:
+                    subprocess.check_output(['latexmk', '-c', filepath],
+                                            stderr=subprocess.STDOUT,
+                                            **check_output_kwargs)
+                except (OSError, IOError, subprocess.CalledProcessError):
                     # Otherwise just remove some file extensions.
                     extensions = ['aux', 'log', 'out', 'fls',
                                   'fdb_latexmk']
 
                     for ext in extensions:
                         try:
-                            os.remove(basename + '.' + ext)
+                            os.remove(filepath + '.' + ext)
                         except (OSError, IOError) as e:
                             # Use FileNotFoundError when python 2 is dropped
                             if e.errno != errno.ENOENT:
@@ -267,7 +287,7 @@ class Document(Environment):
                 rm_temp_dir()
 
             if clean_tex:
-                os.remove(basename + '.tex')  # Remove generated tex file
+                os.remove(filepath + '.tex')  # Remove generated tex file
 
             # Compilation has finished, so no further compilers have to be
             # tried
@@ -276,12 +296,13 @@ class Document(Environment):
         else:
             # Notify user that none of the compilers worked.
             raise(CompilerError(
-                'No LaTex compiler was found\n' +
-                'Either specify a LaTex compiler ' +
+                'No LaTex compiler was found\n'
+                'Either specify a LaTex compiler '
                 'or make sure you have latexmk or pdfLaTex installed.'
             ))
 
-        os.chdir(cur_dir)
+        if not python_cwd_available:
+            os.chdir(cur_dir)
 
     def _select_filepath(self, filepath):
         """Make a choice between ``filepath`` and ``self.default_filepath``.