+++ pylatex/tikz.py
@@ -237,7 +237,7 @@ class TikZNode(TikZObject):
             ret_str.append('({})'.format(self.handle))
 
         if self._node_position is not None:
-            ret_str.append('at {}'.format(str(self._position)))
+            ret_str.append('at {}'.format(str(self._node_position)))
 
         if self._node_text is not None:
             ret_str.append('{{{text}}};'.format(text=self._node_text))