+++ pylatex/utils.py
@@ -29,12 +29,7 @@ _latex_special_chars = {
     ']': r'{]}',
 }
 
-_tmp_path = os.path.abspath(
-    os.path.join(
-        tempfile.gettempdir(),
-        "pylatex"
-    )
-)
+_tmp_path = None
 
 
 def _is_iterable(element):
@@ -106,10 +101,15 @@ def fix_filename(path):
     Latex has problems if there are one or more points in the filename, thus
     'abc.def.jpg' will be changed to '{abc.def}.jpg'
 
+    Windows gets angry about the curly braces that resolve the above issue on
+    linux Latex distributions. MikTeX however, has no qualms about multiple
+    dots in the filename so the behavior is different for posix vs nt when the
+    length of file_parts is greater than two.
+
     Args
     ----
     filename : str
-        The filen name to be changed.
+        The file name to be changed.
 
     Returns
     -------
@@ -134,7 +134,7 @@ def fix_filename(path):
     filename = path_parts[-1]
     file_parts = filename.split('.')
 
-    if len(file_parts) > 2:
+    if os.name == 'posix' and len(file_parts) > 2:
         filename = '{' + '.'.join(file_parts[0:-1]) + '}.' + file_parts[-1]
 
     dir_parts.append(filename)
@@ -326,10 +326,6 @@ def verbatim(s, *, delimiter='|'):
 def make_temp_dir():
     """Create a temporary directory if it doesn't exist.
 
-    Directories created by this functionn follow the format specified
-    by ``_tmp_path`` and are a pylatex subdirectory within
-    a standard ``tempfile`` tempdir.
-
     Returns
     -------
     str
@@ -341,13 +337,16 @@ def make_temp_dir():
     '/var/folders/g9/ct5f3_r52c37rbls5_9nc_qc0000gn/T/pylatex'
     """
 
-    if not os.path.exists(_tmp_path):
-        os.makedirs(_tmp_path)
+    global _tmp_path
+    if not _tmp_path:
+        _tmp_path = tempfile.mkdtemp(prefix="pylatex-tmp.")
     return _tmp_path
 
 
 def rm_temp_dir():
     """Remove the temporary directory specified in ``_tmp_path``."""
 
-    if os.path.exists(_tmp_path):
+    global _tmp_path
+    if _tmp_path:
         shutil.rmtree(_tmp_path)
+        _tmp_path = None