+++ testall.sh
@@ -44,10 +44,10 @@ done
 
 
 # Run the examples and tests
-python_version=$($python --version |& sed 's|Python \(.\).*|\1|g')
+python_version=$($python --version |& sed 's|Python \(.\).*|\1|g' | head -n 1)
 
 # Run the examples and tests
-python_version_long=$($python --version |& sed 's|Python \(.*\)|\1|g')
+python_version_long=$($python --version |& sed 's|Python \(.*\)|\1|g' | head -n 1)
 
 if [ "$python_version" = '3' ]; then
     # Check code guidelines
@@ -66,7 +66,7 @@ else
 fi
 
 echo -e '\e[32mTesting tests directory\e[0m'
-if ! $python "$(which nosetests)" --with-coverage tests/*; then
+if ! $python "$(command -v pytest)" --cov=pylatex tests/*; then
     exit 1
 fi
 mv .coverage{,.tests}
@@ -77,9 +77,9 @@ fi
 
 
 count=0
-for f in $main_folder/examples/*.py; do
+for f in "$main_folder"/examples/*.py; do
     echo -e '\e[32mTesting '"$f"'\e[0m'
-    if ! $python "$(which coverage)" run "$f"; then
+    if ! $python "$(command -v coverage)" run "$f"; then
         exit 1
     fi
     ((count ++))
@@ -100,7 +100,7 @@ if [[ "$nodoc" != 'TRUE' && "$python_version" == "3" && "$python_version_long" !
     ./create_doc_files.sh -p "$python"
     make clean
     set +e
-    if ! $python "$(which sphinx-build)" -b html -d build/doctrees/ source build/html -nW; then
+    if ! $python "$(command -v sphinx-build)" -b html -d build/doctrees/ source build/html -nW; then
         exit 1
     fi
 fi