+++ docs/source/changelog.rst
@@ -14,6 +14,59 @@ This version might not be stable, but to install it use::
 
     pip install git+https://github.com/JelteF/PyLaTeX.git
 
+1.4.1_ - `docs <../v1.4.1/>`__ - 2020-10-18
+-------------------------------------------
+
+Fixed
+~~~~~
+
+- Fixes filename generation with dots in the final filename on Windows.
+- Fixes regression in 1.4.0 where empty ``geometry_options`` would throw an
+  error.
+
+1.4.0_ - `docs <../v1.4.0/>`__ - 2020-09-16
+-------------------------------------------
+
+Added
+~~~~~
+- Add ``Fragment`` class which is a ``Container`` without any LaTeX code
+  surrounding its content.
+
+Fixed
+~~~~~
+- Escape newlines in ``ContainerCommand``
+- Fix bug where the geometry options were not applied in some cases
+
+1.3.4_ - `docs <../v1.3.4/>`__ - 2020-07-29
+-------------------------------------------
+
+Fixed
+~~~~~
+- Use known working versions for Python 3.5 and lower of ordered-set dependency
+
+1.3.3_ - `docs <../v1.3.3/>`__ - 2020-06-20
+-------------------------------------------
+
+Fixed
+~~~~~
+- The 'at' parameter for TikZNode should now work.
+- Use a different temporary directory per user.
+
+1.3.2_ - `docs <../v1.3.2/>`__ - 2020-05-16
+-------------------------------------------
+
+Fixed
+~~~~~
+- On python 3.6+ support multhreaded use of PyLaTeX, by not calling
+  ``os.chdir``
+
+1.3.1_ - `docs <../v1.3.1/>`__ - 2019-09-26
+-------------------------------------------
+
+Fixed
+~~~~~
+- Make labels/sections with weird characters work
+
 1.3.0_ - `docs <../v1.3.0/>`__ - 2017-05-19
 -------------------------------------------
 
@@ -422,7 +475,13 @@ Fixed
 - Fix package delegation with duplicate packages
 
 
-.. _Unreleased: https://github.com/JelteF/PyLaTeX/compare/v1.3.0...HEAD
+.. _Unreleased: https://github.com/JelteF/PyLaTeX/compare/v1.4.1...HEAD
+.. _1.4.1: https://github.com/JelteF/PyLaTeX/compare/v1.4.0...1.4.1
+.. _1.4.0: https://github.com/JelteF/PyLaTeX/compare/v1.3.4...1.4.0
+.. _1.3.4: https://github.com/JelteF/PyLaTeX/compare/v1.3.3...1.3.4
+.. _1.3.3: https://github.com/JelteF/PyLaTeX/compare/v1.3.2...1.3.3
+.. _1.3.2: https://github.com/JelteF/PyLaTeX/compare/v1.3.1...1.3.2
+.. _1.3.1: https://github.com/JelteF/PyLaTeX/compare/v1.3.0...1.3.1
 .. _1.3.0: https://github.com/JelteF/PyLaTeX/compare/v1.2.1...1.3.0
 .. _1.2.1: https://github.com/JelteF/PyLaTeX/compare/v1.2.0...v1.2.1
 .. _1.2.0: https://github.com/JelteF/PyLaTeX/compare/v1.1.1...v1.2.0